Cleanup test codes
