+++ test/v1_5/integration/Gauge/checkpointToken.test.ts
