+++ test/v1_5/integration/VotingEscrow/depositWithdrawVoting.test.ts
@@ -13,9 +13,6 @@ const MAX_EXAMPLES = 30; // テストの試行回数
 const STATEFUL_STEP_COUNT = 20; // ruleの実行回数
 const YEAR = BigNumber.from(86400 * 365);
 const WEEK = BigNumber.from(86400 * 7);
-const two_to_the_256_minus_1 = BigNumber.from("2")
-  .pow(BigNumber.from("256"))
-  .sub(BigNumber.from("1"));
 const ten_to_the_40 = BigNumber.from(
   "10000000000000000000000000000000000000000"
 );
@@ -59,7 +56,7 @@ describe("VotingEscrow", function () {
         ._mintForTesting(accounts[i].address, ten_to_the_40);
       await token
         .connect(accounts[i])
-        .approve(votingEscrow.address, two_to_the_256_minus_1);
+        .approve(votingEscrow.address, ethers.constants.MaxUint256);
     }
 
     //setup
@@ -69,9 +66,13 @@ describe("VotingEscrow", function () {
     });
   });
 
+  afterEach(async () => {
+    await snapshot.restore();
+  });
+
   //--------------------------------------------- functions -----------------------------------------------------------//
 
-  function rdmValue(a): BigNumber {
+  function rdmValue(a: number): BigNumber {
     let rdm = BigNumber.from(Math.floor(Math.random() * a).toString());
     return rdm;
   }
@@ -95,27 +96,22 @@ describe("VotingEscrow", function () {
     unlockTime = timestamp.add(WEEK.mul(stLockDuration)).div(WEEK).mul(WEEK);
 
     if (stValue.eq(0)) {
-      // console.log("--revert: 1");
       await expect(
         votingEscrow.connect(stAccount).createLock(stValue, unlockTime)
       ).to.revertedWith("need non-zero value");
     } else if (votingBalances[stAccountN]["value"].gt("0")) {
-      // console.log("--revert: 2");
       await expect(
         votingEscrow.connect(stAccount).createLock(stValue, unlockTime)
       ).to.revertedWith("Withdraw old tokens first");
     } else if (unlockTime.lte(timestamp)) {
-      // console.log("--revert: 3");
       await expect(
         votingEscrow.connect(stAccount).createLock(stValue, unlockTime)
       ).to.revertedWith("Can only lock until time in the future");
     } else if (unlockTime.gte(timestamp.add(YEAR.mul("4")))) {
-      // console.log("--revert: 4");
       await expect(
         votingEscrow.connect(stAccount).createLock(stValue, unlockTime)
       ).to.revertedWith("Voting lock can be 4 years max");
     } else {
-      // console.log("--success, account:", stAccountN);
       const tx = await votingEscrow
         .connect(stAccount)
         .createLock(stValue, unlockTime);
@@ -141,17 +137,14 @@ describe("VotingEscrow", function () {
     let timestamp = BigNumber.from(await time.latest());
 
     if (stValue.eq(0)) {
-      // console.log("--revert: 1");
       await expect(
         votingEscrow.connect(stAccount).increaseAmount(stValue)
       ).to.revertedWith("dev: need non-zero value");
     } else if (votingBalances[stAccountN]["value"].eq("0")) {
-      // console.log("--revert: 2");
       await expect(
         votingEscrow.connect(stAccount).increaseAmount(stValue)
       ).to.revertedWith("No existing lock found");
     } else if (votingBalances[stAccountN]["unlockTime"].lte(timestamp)) {
-      // console.log("--revert: 3");
       await expect(
         votingEscrow.connect(stAccount).increaseAmount(stValue)
       ).to.revertedWith("Cannot add to expired lock. Withdraw");
@@ -159,13 +152,6 @@ describe("VotingEscrow", function () {
       await votingEscrow.connect(stAccount).increaseAmount(stValue);
       votingBalances[stAccountN]["value"] =
         votingBalances[stAccountN]["value"].add(stValue);
-
-      // console.log(
-      //   "--success, account:",
-      //   stAccountN,
-      //   "new balance:",
-      //   votingBalances[stAccountN]["value"].toString()
-      // );
     }
   }
 
@@ -186,27 +172,22 @@ describe("VotingEscrow", function () {
       .mul(WEEK);
 
     if (votingBalances[stAccountN]["unlockTime"].lte(timestamp)) {
-      // console.log("--revert: 1");
       await expect(
         votingEscrow.connect(stAccount).increaseUnlockTime(unlockTime)
       ).to.revertedWith("Lock expired");
     } else if (votingBalances[stAccountN]["value"].eq("0")) {
-      // console.log("--revert: 2");
       await expect(
         votingEscrow.connect(stAccount).increaseUnlockTime(unlockTime)
       ).to.revertedWith("Nothing is locked");
     } else if (votingBalances[stAccountN]["unlockTime"].gte(unlockTime)) {
-      // console.log("--revert: 3");
       await expect(
         votingEscrow.connect(stAccount).increaseUnlockTime(unlockTime)
       ).to.revertedWith("Can only increase lock duration");
     } else if (unlockTime.gt(timestamp.add(YEAR.mul("4")))) {
-      // console.log("--revert: 4");
       await expect(
         votingEscrow.connect(stAccount).increaseUnlockTime(unlockTime)
       ).to.revertedWith("Voting lock can be 4 years max");
     } else {
-      // console.log("--success, account:", stAccountN);
       const tx = await votingEscrow
         .connect(stAccount)
         .increaseUnlockTime(unlockTime);
@@ -233,7 +214,6 @@ describe("VotingEscrow", function () {
         "The lock didn't expire"
       );
     } else {
-      // console.log("--success, account:", stAccountN);
       await votingEscrow.connect(stAccount).withdraw();
       votingBalances[stAccountN]["value"] = BigNumber.from("0");
     }
@@ -256,25 +236,15 @@ describe("VotingEscrow", function () {
     let stSleepDuration = Math.floor(Math.random() * 3) + 1; //1~4
 
     await time.increase(WEEK.mul(stSleepDuration).toNumber());
-
-    if (stSleepDuration == 1) {
-      console.log("Time advanced");
-    } else {
-      console.log("Time advanced");
-    }
   }
 
   async function checkInvariants() {
-    // console.log("=====Invariant checks=====");
-
-    // console.log("invariant_token_balances");
     for (let i = 0; i < ACCOUNT_NUM; i++) {
       expect(await token.balanceOf(accounts[i].address)).to.equal(
         ten_to_the_40.sub(votingBalances[i]["value"])
       );
     }
 
-    // console.log("invariant_escrow_current_balances");
     let total_supply = BigNumber.from("0");
     let timestamp = BigNumber.from(await time.latest());
 
@@ -294,10 +264,8 @@ describe("VotingEscrow", function () {
     }
     expect(await votingEscrow["totalSupply()"]()).to.equal(total_supply);
 
-    // console.log("invariant_historic_balances");
     total_supply = BigNumber.from("0");
     let blocknumber = (await time.latestBlock()) - 4;
-    // console.log(blocknumber);
 
     for (let i = 0; i < ACCOUNT_NUM; i++) {
       total_supply = total_supply.add(
@@ -311,29 +279,23 @@ describe("VotingEscrow", function () {
   }
 
   let func = [
-    "ruleCreateLock",
-    "ruleIncreaseAmount",
-    "ruleIncreaseUnlockTime",
-    "ruleWithdraw",
-    "ruleCheckpoint",
-    "ruleAdvanceTime",
+    ruleCreateLock,
+    ruleIncreaseAmount,
+    ruleIncreaseUnlockTime,
+    ruleWithdraw,
+    ruleCheckpoint,
+    ruleAdvanceTime,
   ];
 
-  describe("test_deposit_withdraw_voting", function () {
-    //set arbitral number of repeats
-    for (let x = 0; x < MAX_EXAMPLES; x++) {
-      it(`Try ${x}`, async () => {
-        let steps = Math.floor(Math.random() * (STATEFUL_STEP_COUNT - 1)) + 1;
-        for (let i = 0; i < steps; i++) {
-          let n = (await rdmValue(func.length)).toNumber();
-          await eval(func[n])();
-          await checkInvariants();
-        }
-      });
-    }
-  });
-
-  afterEach(async () => {
-    await snapshot.restore();
-  });
+  //set arbitral number of repeats
+  for (let x = 0; x < MAX_EXAMPLES; x++) {
+    it(`Try ${x}`, async () => {
+      let steps = Math.floor(Math.random() * (STATEFUL_STEP_COUNT - 1)) + 1;
+      for (let i = 0; i < steps; i++) {
+        let n = (await rdmValue(func.length)).toNumber();
+        await func[n]();
+        await checkInvariants();
+      }
+    });
+  }
 });
