+++ test/v1_5/unitary/FeeDistributor/claimMultipleTokens.test.ts
@@ -9,11 +9,11 @@ import {
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
 import { sendEther, deploySampleSaleTemplate } from "../../../scenarioHelper";
 
-const DAY = 86400;
-const WEEK = DAY * 7;
-const TEMPLATE_NAME = ethers.utils.formatBytes32String("SampleTemplate");
-
 describe("FeeDistributor", () => {
+  const DAY = 86400;
+  const WEEK = DAY * 7;
+  const TEMPLATE_NAME = ethers.utils.formatBytes32String("SampleTemplate");
+
   let snapshot: SnapshotRestorer;
   let alice: SignerWithAddress,
     bob: SignerWithAddress,
