+++ test/v1_5/unitary/YMWK/burn.test.ts
@@ -24,46 +24,44 @@ describe("YMWK", function () {
     await snapshot.restore();
   });
 
-  describe("YMWK EpochTimeSupply", function () {
-    it("test_burn", async function () {
-      const balance: BigNumber = await token.balanceOf(accounts[0].address);
-      const initialSupply: BigNumber = await token.totalSupply();
+  it("test_burn", async function () {
+    const balance: BigNumber = await token.balanceOf(accounts[0].address);
+    const initialSupply: BigNumber = await token.totalSupply();
 
-      await token.connect(accounts[0]).burn(31337);
+    await token.connect(accounts[0]).burn(31337);
 
-      expect(await token.balanceOf(accounts[0].address)).to.equal(
-        balance.sub(31337)
-      );
-      expect(await token.totalSupply()).to.equal(initialSupply.sub(31337));
-    });
+    expect(await token.balanceOf(accounts[0].address)).to.equal(
+      balance.sub(31337)
+    );
+    expect(await token.totalSupply()).to.equal(initialSupply.sub(31337));
+  });
 
-    it("test_burn_not_admin", async function () {
-      const initialSupply: BigNumber = await token.totalSupply();
+  it("test_burn_not_admin", async function () {
+    const initialSupply: BigNumber = await token.totalSupply();
 
-      await token.transfer(accounts[1].address, 1000000);
-      await token.connect(accounts[1]).burn(31337);
+    await token.transfer(accounts[1].address, 1000000);
+    await token.connect(accounts[1]).burn(31337);
 
-      expect(await token.balanceOf(accounts[1].address)).to.equal(
-        1000000 - 31337
-      );
-      expect(await token.totalSupply()).to.equal(initialSupply.sub(31337));
-    });
+    expect(await token.balanceOf(accounts[1].address)).to.equal(
+      1000000 - 31337
+    );
+    expect(await token.totalSupply()).to.equal(initialSupply.sub(31337));
+  });
 
-    it("test_burn_all", async function () {
-      const initialSupply: BigNumber = await token.totalSupply();
+  it("test_burn_all", async function () {
+    const initialSupply: BigNumber = await token.totalSupply();
 
-      await token.connect(accounts[0]).burn(initialSupply);
+    await token.connect(accounts[0]).burn(initialSupply);
 
-      expect(await token.balanceOf(accounts[0].address)).to.equal(0);
-      expect(await token.totalSupply()).to.equal(0);
-    });
+    expect(await token.balanceOf(accounts[0].address)).to.equal(0);
+    expect(await token.totalSupply()).to.equal(0);
+  });
 
-    it("test_overburn", async function () {
-      const initialSupply: BigNumber = await token.totalSupply();
+  it("test_overburn", async function () {
+    const initialSupply: BigNumber = await token.totalSupply();
 
-      await expect(
-        token.connect(accounts[0]).burn(initialSupply.add(1))
-      ).to.be.revertedWith("ERC20: burn amount exceeds balance");
-    });
+    await expect(
+      token.connect(accounts[0]).burn(initialSupply.add(1))
+    ).to.be.revertedWith("ERC20: burn amount exceeds balance");
   });
 });
