+++ test/v1_5/unitary/YMWK/mint.test.ts
@@ -3,6 +3,7 @@ import { expect } from "chai";
 import {
   takeSnapshot,
   SnapshotRestorer,
+  time,
 } from "@nomicfoundation/hardhat-network-helpers";
 import { BigNumber, Contract } from "ethers";
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
@@ -29,72 +30,61 @@ describe("YMWK", function () {
   afterEach(async () => {
     await snapshot.restore();
   });
-  describe("YMWK Mint", function () {
-    it("test_availableSupply", async function () {
-      const creationTime = await token.startEpochTime();
-      const initialSupply = await token.totalSupply();
-      const rate = await token.rate();
+  it("test_availableSupply", async function () {
+    const creationTime = await token.startEpochTime();
+    const initialSupply = await token.totalSupply();
+    const rate = await token.rate();
 
-      await ethers.provider.send("evm_increaseTime", [week]);
-      await ethers.provider.send("evm_mine", []);
+    await time.increase(week);
 
-      const currentBlock = BigNumber.from(
-        (await ethers.provider.getBlock("latest")).timestamp
-      );
-      // const currentBlock = await ethers.provider.getBlock('latest');
-      const expected = initialSupply.add(
-        currentBlock.sub(creationTime).mul(rate)
-      );
-      expect(await token.availableSupply()).to.equal(expected);
-    });
+    const currentBlock = BigNumber.from(await time.latest());
+    const expected = initialSupply.add(
+      currentBlock.sub(creationTime).mul(rate)
+    );
+    expect(await token.availableSupply()).to.equal(expected);
+  });
 
-    it("test_mint", async function () {
-      await token.setMinter(accounts[0].address);
-      const creationTime = await token.startEpochTime();
-      const initialSupply = await token.totalSupply();
-      const rate = await token.rate();
+  it("test_mint", async function () {
+    await token.setMinter(accounts[0].address);
+    const creationTime = await token.startEpochTime();
+    const initialSupply = await token.totalSupply();
+    const rate = await token.rate();
 
-      await ethers.provider.send("evm_increaseTime", [week]);
+    await ethers.provider.send("evm_increaseTime", [week]);
 
-      const currentTime = BigNumber.from(
-        (await ethers.provider.getBlock("latest")).timestamp
-      );
-      const amount = currentTime.sub(creationTime).mul(rate);
-      await token.mint(accounts[1].address, amount);
+    const currentTime = BigNumber.from(await time.latest());
+    const amount = currentTime.sub(creationTime).mul(rate);
+    await token.mint(accounts[1].address, amount);
 
-      expect(await token.balanceOf(accounts[1].address)).to.equal(amount);
-      expect(await token.totalSupply()).to.equal(initialSupply.add(amount));
-    });
+    expect(await token.balanceOf(accounts[1].address)).to.equal(amount);
+    expect(await token.totalSupply()).to.equal(initialSupply.add(amount));
+  });
 
-    it("test_overmint", async function () {
-      await token.setMinter(accounts[0].address);
-      const creationTime = await token.startEpochTime();
-      const rate = await token.rate();
+  it("test_overmint", async function () {
+    await token.setMinter(accounts[0].address);
+    const creationTime = await token.startEpochTime();
+    const rate = await token.rate();
 
-      await ethers.provider.send("evm_increaseTime", [week]);
-      await ethers.provider.send("evm_mine", []);
+    await time.increase(week);
 
-      const currentTime = BigNumber.from(
-        (await ethers.provider.getBlock("latest")).timestamp
-      );
-      const amount = currentTime.sub(creationTime).add(2).mul(rate);
-      await expect(token.mint(accounts[1].address, amount)).to.be.revertedWith(
-        "dev: exceeds allowable mint amount"
-      );
-    });
+    const currentTime = BigNumber.from(await time.latest());
+    const amount = currentTime.sub(creationTime).add(2).mul(rate);
+    await expect(token.mint(accounts[1].address, amount)).to.be.revertedWith(
+      "dev: exceeds allowable mint amount"
+    );
+  });
 
-    it("test_minter_only", async function () {
-      await token.setMinter(accounts[0].address);
-      await expect(
-        token.connect(accounts[1]).mint(accounts[1].address, 0)
-      ).to.be.revertedWith("dev: minter only");
-    });
+  it("test_minter_only", async function () {
+    await token.setMinter(accounts[0].address);
+    await expect(
+      token.connect(accounts[1]).mint(accounts[1].address, 0)
+    ).to.be.revertedWith("dev: minter only");
+  });
 
-    it("test_zero_address", async function () {
-      await token.setMinter(accounts[0].address);
-      await expect(token.mint(ZERO_ADDRESS, 0)).to.be.revertedWith(
-        "dev: zero address"
-      );
-    });
+  it("test_zero_address", async function () {
+    await token.setMinter(accounts[0].address);
+    await expect(token.mint(ZERO_ADDRESS, 0)).to.be.revertedWith(
+      "dev: zero address"
+    );
   });
 });
